Move character database related stuff to submodule
